requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pyswisseph==2.10.3.2
supabase==2.4.1
pydantic==2.6.4
python-dotenv==1.0.1
